Issue #1555: Remove unnecessary unboxing
Fixes `UnnecessaryUnboxing` inspection violations.

Description:
>Reports "unboxing", e.g. explicit unwrapping of wrapped primitive values. Unboxing is unnecessary under Java 5 and newer, and can be safely removed.
This inspection only reports if the project or module is configured to use a language level of 5.0 or higher.

+++ src/main/java/com/puppycrawl/tools/checkstyle/checks/TrailingCommentCheck.java
@@ -179,7 +179,7 @@ public class TrailingCommentCheck extends AbstractFormatCheck {
             }
             if (!blankLinePattern.matcher(lineBefore).find()
                 && !isLegalComment(comment)) {
-                log(lineNo.intValue(), MSG_KEY);
+                log(lineNo, MSG_KEY);
             }
         }
     }
